Do not use deprecated app=prometheus label (#6)
This commit changes references from the `app=prometheus` label to the
`prometheus=k8s`. This is because prometheus-operator deprecates the
`app` label in 0.48.0 and removes it in 0.50.

As a consequence, the current label filter does not work anymore in OCP
4.10. We are moving to `prometheus=k8s` as this is the one that is
present in OCP 4.6 and later, which are supported releases for the GPU
Operator.

Relates-to: https://github.com/openshift/cluster-monitoring-operator/pull/1388